requirements.txt
python-telegram-bot==13.13
pandas==1.4.3
requests==2.28.1
lxml==4.9.0
python-dotenv==0.20.0